Moving data to _data - 2.
